:truck: Rename provider module to store
